% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HCLUST.R
\name{HCLUST}
\alias{HCLUST}
\title{Hierarchical clustering}
\usage{
HCLUST(data.mat, K)
}
\arguments{
\item{data.mat}{input data matrix, [number of samples] x  [number of features]}

\item{K}{number of clusters}
}
\value{
Vector of cluster assignments, [number of samples]
}
\description{
Agglomerative hierarchical clustering on a given data matrix with a given
number of clusters. Uses single-linkage method.
}
\examples{
HCLUST(iris[,1:4],3)
}
